requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
decorator==4.4.2
fireREST==1.0.4
flake8==3.8.4
idna==2.10
mccabe==0.6.1
packaging==20.9
paramiko==2.7.2
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-decouple==3.4
python-dotenv==0.15.0
requests==2.25.1
retry==0.9.2
simplejson==3.17.2
six==1.15.0
urllib3==1.26.3
